+++ src/header/Stash Scene Tagger Draft Submit.user.js
@@ -2,7 +2,7 @@
 // @name        Stash Scene Tagger Draft Submit
 // @namespace   %NAMESPACE%
 // @description Adds button to Scene Tagger to submit draft to stashdb
-// @version     0.1.0
+// @version     0.1.1
 // @author      7dJx1qP
 // @match       %MATCHURL%
 // @grant       unsafeWindow